播放列表的更新已解决

+++ app/src/main/java/com/leon/biuvideo/beans/musicBeans/MusicPlayList.java
@@ -6,7 +6,7 @@ public class MusicPlayList {
     }
 
     public int id;
-    public int sid;
+    public long sid;
     public String musicName;
     public String author;
 }
